Address comments: 1. Fontsize as float not uint 2. pass full KBS in, gives access to all settings and events 3. Move event handling into PlainTextPageSource, so all consumers will benefit. 4. rename mMessages to better show purpose

+++ src/app/app-common/include/OpenKneeboard/TextSettings.h
@@ -23,7 +23,7 @@
 
 namespace OpenKneeboard {
 struct TextSettings final {
-  uint32_t mFontSize = 12;
+  float mFontSize = 12.f;
 
   constexpr auto operator<=>(const TextSettings&) const = default;
 };
